backend/requirements.txt

fastapi==0.110.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pyjwt==2.8.1
python-dotenv==1.0.0
pydantic[email]==2.5.3
python-multipart==0.0.6
httpx==0.25.2